updated certa example

+++ certa/eval.py
@@ -39,9 +39,9 @@ def expl_eval(class_to_explain, explanation_attributes, e_score, lsource, l_reco
             class_probability = prediction[class_to_explain]
             drop = class_probability - modified_tuple_prediction[class_to_explain]
             n_drop = (class_probability - modified_tuple_prediction[class_to_explain]) / class_probability
-            impact = flip or (modified_tuple_prediction[class_to_explain] < 0.5 * class_probability)
-            n_impact = flip or (
-                    modified_tuple_prediction[class_to_explain] / class_probability < 0.5 * class_probability)
+            impact = int(flip or (modified_tuple_prediction[class_to_explain] < 0.5 * class_probability))
+            n_impact = int(flip or (
+                    modified_tuple_prediction[class_to_explain] / class_probability < 0.5 * class_probability))
             ids = str(l_record['id']) + '-' + str(r_record['id'])
             new_row = {'match': class_to_explain, 'e_score': e_score, 'drop': drop, 'n_drop': n_drop,
                        'perturb': perturb, 'flip': flip, 'impact': impact, 'n_impact': n_impact,
